[FIX]: 🔧 usecase에 final keyword 추가
