--- PopPool/PopPool/Domain/UseCase/SaveTokenUseCaseImpl.swift
@@ -1,21 +1,0 @@
-//
-//  SaveTokenUseCaseImpl.swift
-//  PopPool
-//
-//  Created by Porori on 6/19/24.
-//
-
-import Foundation
-import RxSwift
-
-final class SaveTokenUseCaseImpl: LocalSaveUseCase {
-    var repository: LocalDBRepository
-    
-    init(repository: LocalDBRepository) {
-        self.repository = repository
-    }
-    
-    func save(key: String, value: String, to databaseType: String) -> Completable {
-        repository.save(key: key, value: value, to: databaseType)
-    }
-}
